refactor: rename the ssh_cmd module to ssh
The module wrapped the ssh binary once and it no longer does, so the
cmd suffix described something that had gone away. The client type
was named after the russh crate rather than its job and is now
SshClient, in a file that no longer shadows that crate.
