+++ src/ssh.rs
@@ -1,9 +1,9 @@
 mod port_checker;
-mod russh;
 mod sftp_path;
+mod ssh_client;
 mod ssh_key_generator;
 
 pub use port_checker::PortChecker;
-pub use russh::Russh;
 pub use sftp_path::SftpPath;
+pub use ssh_client::SshClient;
 pub use ssh_key_generator::SshKeyGenerator;
